requirements.txt
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
anyio==4.8.0
blessed==1.20.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
Cython==3.0.11
et_xmlfile==2.0.0
eval_type_backport==0.2.0
fastapi==0.115.8
filelock==3.16.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fpdf==1.7.2
fsspec==2024.10.0
gensim==4.3.3
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
imageio==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==3.0.2
milvus-lite==2.4.11
milvus-model==0.2.12
mpmath==1.3.0
networkx==3.4.2
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
onnxruntime==1.20.0
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
plotly==5.19.0
protobuf==5.28.3
psutil==6.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pymilvus==2.5.4
PyMuPDF==1.24.12
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyvi==0.1.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.2.1
simsimd==6.0.6
six==1.16.0
sklearn-crfsuite==0.5.0
smart-open==7.0.5
sniffio==1.3.1
soundfile==0.13.0
SQLAlchemy==2.0.38
starlette==0.45.3
stringzilla==3.10.10
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.6
transformers==4.46.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.2.1
